CHORE: Implemented TS
